apps/backend/pyproject.toml
[project]
name = "refactor-agent"
version = "0.1.0"
description = "Agentic refactoring"
readme = "../../README.md"
authors = [
    { name = "Thomas Decloedt", email = "thomas.r.decloedt@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.59.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.78.0,<0.80.0",  # >=0.78 needed for Claude 4.6; <0.80 avoids UserLocation import break in pydantic-ai (PR #4345 pending)
    "langfuse",
    "python-dotenv",
    "libcst>=1.0.0",
    "fastmcp>=2.14.0,<3",
    "a2a-sdk>=0.3.0",
    "uvicorn[standard]>=0.32.0",
    "starlette>=0.38.0",
    "websockets>=14.0",
    "watchdog>=4.0",
    "chainlit",
    "asyncpg",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "sentry-sdk>=2.0.0",
    "google-cloud-firestore>=2.19.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "pyright",
    "chainlit",
    "types-PyYAML",
    "vulture>=2.11",
    "flake8>=7.0",
    "flake8-deprecated>=2.3",
    "pre-commit>=4.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

# ---------------------------------------------------------------------------
# Mypy – static type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "stubs"
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
warn_unused_ignores = true  # Error on unnecessary # type: ignore comments
extra_checks = true
enable_error_code = [
    "ignore-without-code",
    "redundant-cast",
    "truthy-bool",
    "truthy-iterable",
    "unused-awaitable",
]
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# ---------------------------------------------------------------------------
# Pyright – static type checking (runs alongside mypy; catches different edge cases)
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
stubPath = "stubs"
# Warning: mypy needs some type: ignore that pyright considers unnecessary (different implementations).
reportUnnecessaryTypeIgnoreComment = "warning"
# Third-party packages without stubs: downgrade to warning (matches mypy ignore_missing_imports)
reportMissingImports = "warning"
reportMissingModuleSource = "warning"

# Per-path overrides for modules with known third-party/framework typing gaps (matches mypy overrides)
[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/engine/python"
reportAttributeAccessIssue = "warning"
reportIncompatibleMethodOverride = "warning"

[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/observability"
reportArgumentType = "warning"

[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/agent"
reportArgumentType = "warning"

[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/schedule"
reportArgumentType = "warning"

[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/sync"
reportArgumentType = "warning"

[[tool.pyright.executionEnvironments]]
root = "src/refactor_agent/ui"
reportArgumentType = "warning"
reportIndexIssue = "warning"

[[tool.mypy.overrides]]
# pydantic_ai._agent_graph has End, ModelRequestNode at runtime but not in __all__.
module = ["pydantic_ai._agent_graph"]
ignore_missing_imports = true
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
# Backend + method_logging: ASGI app typing (object/callable), scope dicts.
module = ["refactor_agent.backend.*", "refactor_agent.a2a.method_logging"]
disable_error_code = ["operator", "type-arg", "unreachable"]

[[tool.mypy.overrides]]
module = [
    "langfuse.*",
    "pydantic_ai.*",
    "pydantic_ai._utils",
    "dotenv.*",
    "libcst.*",
    "fastmcp.*",
    "a2a.*",
    "chainlit.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Chainlit has no type stubs; strict untyped-call/def checks are unworkable.
module = ["refactor_agent.ui.*"]
disallow_untyped_calls = false
disallow_untyped_defs = false

[[tool.mypy.overrides]]
# LibCST typed transformer return type is narrower (FlattenSentinel/RemovalSentinel); we return node.
module = ["refactor_agent.engine.python.libcst_engine"]
disable_error_code = ["override"]

# ---------------------------------------------------------------------------
# Vulture – dead code detection
# ---------------------------------------------------------------------------
[tool.vulture]
min_confidence = 80
paths = ["src", "tests", "scripts"]
ignore_names = ["_", "logger", "app", "router", "client", "rest", "current_user", "UserLocation", "FieldInfo"]
exclude = ["*/node_modules/*", "*/__pycache__/*", "*.pyc"]

# ---------------------------------------------------------------------------
# Flake8 (deprecation only; ruff handles other linting)
# ---------------------------------------------------------------------------
[tool.flake8]
select = ["D"]
extend-ignore = ["E", "F", "W"]
exclude = [".git", "__pycache__", "node_modules", ".venv", "venv"]
max-line-length = 88

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
strict_markers = true
filterwarnings = ["error"]
